Remove syntastic and add ale
I've had some problems with syntastic, namely:

* difficulty getting it to find the `eslint` executable on my new Stride
    laptop. Hopefully ale won't have this problem.
* it is super slow when I have several files open, because it
    processings all of them on save. Ale is asynchronous and lints files
    while I edit them, so hopefully it'll perform much better.

---

This commit also adds GraphQL tags explicitly. I just used the default
options, which I shouldn't have needed to specify at all, but I was
getting an error stating that the tag variable was undefined, and
setting it removed the issue. Hopefully in the future I'll figure out
why the default wasn't loading from the package and not have to specify
it in `vimrc`.